requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
albumentations==1.3.0
apex==0.1
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blessed==1.20.0
cachetools==5.3.0
certifi==2023.7.22
charset-normalizer==3.1.0
click==8.1.7
clip==1.0
cmake==3.26.3
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
datasets==2.14.6
debugpy==1.6.7
decorator==5.1.1
dill==0.3.7
docker-pycreds==0.4.0
einops==0.6.0
executing==1.2.0
filelock==3.11.0
fire==0.5.0
fonttools==4.39.3
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.1.1
gdown==4.7.1
gitdb==4.0.11
GitPython==3.1.40
gpustat==1.1
h5py==3.9.0
huggingface-hub==0.18.0
idna==3.4
imageio==2.27.0
imagenetv2-pytorch==0.1
importlib-metadata==6.4.1
importlib-resources==5.12.0
ipdb==0.13.13
ipykernel==6.22.0
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
lazy_loader==0.2
lit==16.0.1
littleutils==0.2.2
loralib==0.1.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpi4py==3.1.4
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
munch==2.5.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==11.525.112
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
nvitop==1.1.2
ogb==1.3.6
open-clip-torch==2.23.0
opencv-python-headless==4.7.0.72
outdated==0.2.2
packaging==23.1
pandas==2.0.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pip==23.3.1
platformdirs==3.2.0
prettytable==3.7.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.0
Pygments==2.15.0
PyHessian==0.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
qudida==0.0.4
regex==2023.10.3
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
safetensors==0.3.0
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
sconf==0.2.5
seaborn==0.12.2
sentencepiece==0.1.99
sentry-sdk==1.32.0
setproctitle==1.3.3
setuptools==65.6.3
six==1.16.0
smmap==5.0.1
soupsieve==2.4.1
stack-data==0.6.2
sympy==1.11.1
tabulate==0.9.0
tensorboardX==2.6
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.8.19.dev0
tomli==2.0.1
torch==2.0.0
torchaudio==2.0.1
torchmetrics==0.11.4
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wandb==0.15.12
wcwidth==0.2.6
wheel==0.38.4
wilds==2.0.0
xxhash==3.4.1
yarl==1.9.2
zipp==3.15.0
